+++ mlir/lib/Compiler/CompilerPipeline.cpp
@@ -13,6 +13,7 @@
 #include "mlir/Conversion/QCOToQC/QCOToQC.h"
 #include "mlir/Conversion/QCToQCO/QCToQCO.h"
 #include "mlir/Conversion/QCToQIR/QCToQIR.h"
+#include "mlir/Dialect/QCO/Transforms/Passes.h"
 #include "mlir/Support/PrettyPrinting.h"
 
 #include <llvm/ADT/StringRef.h>
@@ -142,8 +143,11 @@ QuantumCompilerPipeline::runPipeline(ModuleOp module,
     }
   }
   // Stage 5: Optimization passes
-  // TODO: Add optimization passes
-  if (failed(runStage([&](PassManager& pm) { addCleanupPasses(pm); }))) {
+  if (failed(runStage([&](PassManager& pm) {
+        if (config_.hadamardLifting) {
+          pm.addPass(mlir::qco::createHadamardLifting());
+        }
+      }))) {
     return failure();
   }
   if (record != nullptr && config_.recordIntermediates) {